Did some dB, viewAll, working on runCode
Able to check if there is a file in the CorrectCode folder, if so link it to the dB if not ask the User again. Included an NA option if there are topics that do not have any code to it.

Fixed the selection since it kept asking the user for another topic after selecting N.

Included runner(), which is now able to run code from other files.

TODO: still need to utilize the runCode class to make it either run the correctCode file or the testFile that the user will implement.

Thinking about including a JForm?